Fix wrong pagenumber and no-mode-line scroll
When the next/previous page has been loaded (by scrolling over the edge
of a page to the next/previous page), then the display buffer will create a new
buffer at that next/previous page. So we first have to set the new buffer to the
page of the current display.